requirements.txt
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
ecdsa==0.17.0
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
peewee==3.15.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.18.1
watchfiles==0.15.0
websockets==10.3
